Add NewEntry model
This model will be used for new entries that are to be saved to the
server. The sync function has been overridden so it passes everything
the server should need to process the request.

We turn on Backbone.emulateHTTP here, so all modifying actions will be
sent as HTTP POSTs.