Add an EntriesCollection + some refactoring
This collection is used to keep track of what entries are on the page.
It listens to the EntriesQueue sync event, updates itself accordingly -
and removes entries that it's processed from the EntriesQueue.

EntriesView now listens to add events on EntriesCollection.

EntryView is now responsible for updating its own time.