Make sure we don't reset the collection on fetch
Now we listen to the sync event, instead of the reset event - the reset
event is only fired when the entire collection has been replaced. This
isn't what we want, we just want to update the queue.

Fetch should now ensure that the reset event isn't fired, and we don't
remove anything we shouldn't from the queue.

Fetch will also update the attributes of an entry with the same ID.