Generate our own fetched event, instead of using sync or reset
Since Backbone 0.9.2's fetch does not emit an sync event, and we don't
want to reset a collection as it will replace all the contents of the
EntriesQueue, we now emit our own 'fetched' event.

Also, we patch Backbone 0.9.2's Collection.add to implement the merge
option present in later versions. This will update the entry with the
same ID we get back from the server, instead of just discarding it as as
a duplicate. This is useful when we have a new entry in the queue, and
that entry is then updated before the nag has been clicked.